Добработка авторизации, регистрации, изменения аккаунта
