+++ src/components/Profile/Profile.css
@@ -146,6 +146,10 @@
     color: #3ddc84;
 }
 
+.profile__message_error {
+    color: #EE3465;
+}
+
 @media screen and (max-width: 469px) {
     .profile__account {}
 }
